refactor(GameTrialCount): GameTrial -> GameTrialCount로 클래스 이름 변경
