--- src/main/java/domain/GameTrial.java
@@ -1,18 +1,0 @@
-package domain;
-
-public class GameTrial {
-
-    int gameTrial;
-
-    public GameTrial(int gameTrial) {
-        this.gameTrial = gameTrial;
-    }
-
-    public boolean isGreaterThan(int otherCount) {
-        return gameTrial > otherCount;
-    }
-
-    public void reduce() {
-        gameTrial--;
-    }
-}
